requirements.txt
anyio==3.6.1
astroid==2.11.7
autopep8==1.6.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
dill==0.3.5.1
fastapi==0.79.0
filelock==3.7.1
h11==0.13.0
httptools==0.4.0
huggingface-hub==0.8.1
idna==3.3
isort==5.10.1
Jinja2==3.1.2
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
numpy==1.23.1
packaging==21.3
Pillow==9.2.0
platformdirs==2.5.2
pycodestyle==2.8.0
pydantic==1.9.1
pylint==2.14.5
pyparsing==3.0.9
python-dotenv==0.20.0
python-magic==0.4.27
python-multipart==0.0.5
PyYAML==6.0
regex==2022.7.25
requests==2.28.1
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
tokenizers==0.12.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.1
torch==1.12.0
torchaudio==0.12.0
torchvision==0.13.0
tqdm==4.64.0
transformers==4.20.1
typing_extensions==4.3.0
urllib3==1.26.11
uvicorn==0.18.2
watchfiles==0.16.0
websockets==10.3
wrapt==1.14.1
